refactor: Remove DoctorServices and PatientClinicServices and their dependencies
