--- back-end/SkinCancer.Services/DoctorServices/IDoctorService.cs
@@ -1,24 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
-using SkinCancer.Entities.AuthModels;
-using SkinCancer.Entities.ModelsDtos.DoctorDtos;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SkinCancer.Services.DoctorServices
-{
-    public interface IDoctorService
-    {
-        /*Task<DoctorClinicDto> CreateClinicAsync(DoctorClinicDto dto);
-
-        Task<ProcessResult> DeleteClinicAsync(int id);
-
-        Task<ProcessResult> GetClinicByName(string name);
-
-        Task<ProcessResult> GetClinicById(int id);*/
-
-    }
-}
